fmt(child_proc): separate file
