+++ Makefile
@@ -13,6 +13,7 @@ SRCS		+=	$(DEBUG_DIR)/put.c
 
 EXEC_DIR	:=	exec
 SRCS		+=	$(EXEC_DIR)/check_command.c \
+				$(EXEC_DIR)/child_pipes.c \
 				$(EXEC_DIR)/child_process.c \
 				$(EXEC_DIR)/exec.c \
 				$(EXEC_DIR)/parent_pipes.c \
